Add lint config files for several linters.
This change adds config files for coffeelint, scss-lint, and rubocop.
Because these linters allow configuration files to be specified from
the command line, I've updated atom-lint to allow this to be specified
in the config.cson file.

The config files can be access from `atom-lint.<linter>.config`. The
value for these keys should be the path to the file.